requirements.txt
PyYAML>=6.0.1
httpx>=0.27.0
loguru>=0.7.2
async-tkinter-loop>=0.9.3
pydantic>=2.7.0
plyer>=2.1.0
